requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
boto3>=1.34.0
rich>=13.7.0
click>=8.1.7
Pillow>=10.0.0
tkterm>=0.0.0b2
pyte>=0.8.0
ujenkins>=0.11.1
pycryptodome>=3.20.0
selenium>=4.0.0
webdriver-manager>=4.0.0
PyGithub>=1.59.0
tksheet>=7.0.0
html2text>=2020.1.16
pygments>=2.17.0
psycopg2-binary>=2.9
sqlalchemy>=2.0
psutil>=5.9.0
pystray>=0.19.0
